Move parent component to frontend folder
