+++ frontend/components/index.js
@@ -1,0 +1,3 @@
+export * from './Counter';
+export * from './Foods';
+export * from './Image';
